Crate async_dnssd

Asynchronous wrapper for DNS-SD C libraries

Interesting entry points:

Browse for available services

Create Connection to register records with

Enumerate domains that are recommended for registration or browsing

Query for an arbitrary DNS record

Register a service

Add a record to a registered service

Register record

Find hostname and port (and more) for a service

Also the following things might be interesting:

Purge record from cache

Construct full name

Stream timeouts

Porting from dnssd C API

C API	functionality in this crate

DNSServiceAddRecord	Registration::add_record, Register::add_record

DNSServiceBrowse	browse

DNSServiceConstructFullName	FullName::construct

DNSServiceCreateConnection	connect

DNSServiceEnumerateDomains	enumerate_domains

DNSServiceQueryRecord	query_record

DNSServiceReconfirmRecord	reconfirm_record

DNSServiceRegister	register

DNSServiceRegisterRecord	Connection::register_record

DNSServiceResolve	resolve

DNSServiceUpdateRecord	Record::update_record, RegisterRecord::update_record

The following functions are called automatically when needed:

DNSServiceProcessResult driving callbacks (event loop)

DNSServiceRefDeallocate called when dropping various resource handles

DNSServiceRefSockFD used for integration with tokio (event loop)

DNSServiceRemoveRecord called when dropping Record

The TXTRecord* “TXT Record Construction Functions” are not wrapped; TxtRecord provides a native rust implementation with similar functionality.

Structs

Browse

Pending browse request

BrowseData

Optional data when browsing for a service; either use its default value or customize it like:

BrowseResult

Browse result

BrowsedFlags

Flags for BrowseResult

Class

DNS CLASS

Connection

Connection to register records with

EnumerateDomains

Pending domain enumeration

EnumerateResult

Domain enumeration result

EnumeratedFlags

Flags for EnumerateDomains

FullName

Full name consiting of (up to) three parts

InterfaceIndex

Network interface index

QueriedRecordFlags

Flags for QueryRecordResult

QueryRecord

Pending query

QueryRecordData

Optional data when querying for a record; either use its default value or customize it like:

QueryRecordFlags

Flags used to query for a record

QueryRecordResult

Query result

Record

A successful record registration

Register

Pending registration

RegisterData

Optional data when registering a service; either use its default value or customize it like:

RegisterFlags

Flags used to register service

RegisterRecord

Pending record registration

RegisterRecordData

Optional data when registering a record; either use its default value or customize it like:

RegisterRecordFlags

Flags used to register a record

RegisterResult

Service registration result

Registration

Successful registration

Resolve

Pending resolve request

ResolveHost

Pending resolve

ResolveHostData

Optional data when querying for a record; either use its default value or customize it like:

ResolveHostResult

Resolve host result

ResolveResult

Resolve result

ResolvedFlags

Flags for ResolveResult

ResolvedHostFlags

Flags for ResolveHostResult

TimeoutStream

Add a timeout to a stream; each time an item is received the timer is reset

TxtRecord

Key-Value container that uses DNS TXT RDATA as representation

TxtRecordIter

Iterator for entries in TxtRecord

Type

DNS (RR)TYPE

Enums

Enumerate

Whether to enumerate domains which are browsed or domains for which registrations can be made.

Error

API Error

Interface

Network interface

ScopedSocketAddr

IP address with port and “scope id” (even for IPv4)

TxtRecordError

Error returned when inserting new entries failed

Constants

MAX_DOMAIN_NAME

Maximum length of full name including trailing dot and terminating NULL

Traits

StreamTimeoutExt

Stream extension to simplify building TimeoutStream

Functions

browse

Browse for available services

browse_extended

Browse for available services

connect

Create Connection to register records with

enumerate_domains

Enumerate domains that are recommended for registration or browsing

query_record

Query for an arbitrary DNS record

query_record_extended

Query for an arbitrary DNS record

reconfirm_record

Purge record from cache

register

Register a service

register_extended

Register a service

resolve

Find hostname and port (and more) for a service

resolve_host_extended

Resolves hostname (with passed port) to stream of ScopedSocketAddr.